Separate Settings from SettingsItem
