+++ src/bezel/settings.ts
@@ -1,5 +1,4 @@
 import * as vscode from 'vscode';
-import { BuiltInCommands } from '../constants';
 import { Container } from '../container';
 import { ComponentConfiguration } from './configuration';
 
@@ -29,40 +28,21 @@ function getConfigurationProperties(prefix: string): Map<string, ConfigurationPr
 }
 
 export abstract class Settings<T extends ComponentConfiguration>
-  extends vscode.TreeItem
   implements vscode.Disposable
 {
   protected disposables: vscode.Disposable[] = [];
   private cfg: Promise<T> | undefined;
-  private props: Map<string, ConfigurationProperty>;
+  public readonly props: Map<string, ConfigurationProperty>;
   private _onDidConfigurationChange: vscode.EventEmitter<T> = new vscode.EventEmitter();
   public onDidConfigurationChange: vscode.Event<T> = this._onDidConfigurationChange.event;
   private _onDidConfigurationError: vscode.EventEmitter<Error> = new vscode.EventEmitter();
   public onDidConfigurationError: vscode.Event<Error> = this._onDidConfigurationError.event;
 
   constructor(public readonly section: string) {
-    super('Settings');
     this.props = getConfigurationProperties(section);
-    this.description = section;
-    this.iconPath = new vscode.ThemeIcon('gear');
-    this.tooltip = new vscode.MarkdownString(
-      `### Settings for "${section}"
-      
-      Click to open the VSCode settings and update the configuration items as desired.
-
-      Changes should be reflected automatically, you should not need to reload the window.
-      `
-    );
-    this.collapsibleState = vscode.TreeItemCollapsibleState.Collapsed;
-    this.command = {
-      title: 'Edit settings',
-      command: BuiltInCommands.OpenSettings,
-      arguments: [section],
-    };
 
     this.disposables.push(this._onDidConfigurationChange);
     this.disposables.push(this._onDidConfigurationError);
-
     this.disposables.push(
       vscode.workspace.onDidChangeConfiguration(async e => {
         if (e.affectsConfiguration(section)) {
@@ -77,7 +57,7 @@ export abstract class Settings<T extends ComponentConfiguration>
     try {
       const config = vscode.workspace.getConfiguration(section);
       if (!config) {
-        throw new Error(`error: configuration section "${section}" not found.`);
+        throw new Error(`configuration section "${section}" not found.`);
       }
       const cfg = await this.configure(config);
 
@@ -88,16 +68,12 @@ export abstract class Settings<T extends ComponentConfiguration>
 
       this.cfg = Promise.resolve(cfg);
       this._onDidConfigurationChange.fire(cfg);
-      return this.cfg;
     } catch (e) {
-      this.iconPath = new vscode.ThemeIcon('warning');
-      this.collapsibleState = vscode.TreeItemCollapsibleState.None;
-
-      const msg = `could not configure "${section}": ${e.message}`;
+      this.cfg = Promise.reject(`could not configure "${section}": ${e.message}`)
       this._onDidConfigurationError.fire(e);
-      this.description = e.message;
-      return (this.cfg = Promise.reject(msg));
     }
+
+    return this.cfg;
   }
 
   protected abstract configure(config: vscode.WorkspaceConfiguration): Promise<T>;
@@ -107,21 +83,6 @@ export abstract class Settings<T extends ComponentConfiguration>
       this.cfg = this.reconfigure(this.section);
     }
     return this.cfg;
-  }
-
-  async getChildren(): Promise<vscode.TreeItem[] | undefined> {
-    return Array.from(this.props.values()).map(p => {
-      const item = new vscode.TreeItem(p.name);
-      item.description = formatValue(p.value, p.default);
-      item.tooltip = p.description;
-      item.iconPath = new vscode.ThemeIcon(getThemeIconNameForPropertyType(p.type));
-      item.command = {
-        title: 'Edit Setting',
-        command: BuiltInCommands.OpenSettings,
-        arguments: [p.key],
-      };
-      return item;
-    });
   }
 
   dispose() {
@@ -137,29 +98,4 @@ interface ConfigurationProperty {
   description: string;
   type: string;
   default?: any;
-}
-
-function getThemeIconNameForPropertyType(type: string): string {
-  switch (type) {
-    case 'string':
-      return 'symbol-string';
-    case 'array':
-      return 'symbol-array';
-    case 'number':
-      return 'symbol-number';
-    case 'boolean':
-      return 'symbol-boolean';
-    default:
-      return 'symbol-property';
-  }
-}
-
-function formatValue(v: any, def?: any): string {
-  if (Array.isArray(v)) {
-    return v.join(' ');
-  }
-  if (typeof v === 'undefined') {
-    return '';
-  }
-  return String(v);
 }
